+++ catkin_ws/src/autoSSAR/CMakeLists.txt
@@ -13,6 +13,8 @@ find_package(catkin REQUIRED COMPONENTS
   std_msgs
   geometry_msgs
   pcl_conversions
+  octomap_msgs
+  octomap
 )
 
 ## System dependencies are found with CMake's conventions
@@ -105,7 +107,7 @@ find_package(catkin REQUIRED COMPONENTS
 ## CATKIN_DEPENDS: catkin_packages dependent projects also need
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
-#  INCLUDE_DIRS include
+ INCLUDE_DIRS include
 #  LIBRARIES autoSSAR
 #  CATKIN_DEPENDS roscpp rospy std_msgs
 #  DEPENDS system_lib
@@ -118,7 +120,7 @@ catkin_package(
 ## Specify additional locations of header files
 ## Your package locations should be listed before other locations
 include_directories(
-# include
+  include
   ${catkin_INCLUDE_DIRS}
 )
 
@@ -150,6 +152,11 @@ add_executable(mapmerge src/mapmerge.cpp)
 ## Specify libraries to link a library or executable target against
 target_link_libraries(mapmerge
   ${catkin_LIBRARIES}
+)
+
+target_include_directories(${PROJECT_NAME}
+  ${catkin_INCLUDE_DIRS}
+  /include
 )
 
 #############
