+++ catkin_ws/src/autoSSAR/include/mapMerge/mapmerge.h
@@ -4,10 +4,13 @@
 
 #include <ros/ros.h>
 #include <string>
+#include <pcl/common/projection_matrix.h>
+#include <pcl/kdtree/kdtree_flann.h>
+#include <pcl_conversions/pcl_conversions.h>
 
 namespace mapMerge
 {
-    void mergeMaps(pcl::PointCloud<pcl::PointXYZ>& cloudMap);
+    void mergeMaps(pcl::PointCloud<pcl::PointXYZ>& map_in, octomap_msgs::Octomap map_out);
 
 
 }
